refactor(user): 리패키징
