+++ src/main/java/com/sprarta/sproutmarket/domain/auth/dto/response/OAuthTokenResponse.java
@@ -1,0 +1,13 @@
+package com.sprarta.sproutmarket.domain.auth.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class OAuthTokenResponse {
+    private String access_token;
+    private String refresh_token;
+    private String token_type;
+    private String expires_in;
+}
